Fix config upgrades
The previous update changed the major version of the config from 1 to 2
and then made it so thatmigration always failed. The intention was that
reconfiguration would somehow automatically fix this, but sadly that
didn't work. With this change the config is automatically migrated to 2.
Just to make this more complicated, the HA API for this is changing in
2024.3, so there are two code paths to handle the version change.

The API key is then checked during initialization and if it doesn't
work, ConfigEntryAuthFailed is raised. This shows up as a repair in
settings which prompts the user to enter an API key.